refactor: split timing-sensitive production modules
